requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
black==21.12b0
cffi==1.15.0
channels==2.4.0
channels-redis==2.4.2
click==8.0.3
constantly==15.1.0
cryptography==36.0.1
daphne==2.5.0
dataclasses==0.8
Django==3.2.10
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
flake8==4.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.2.0
incremental==21.3.0
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==1.7.1
pyOpenSSL==21.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
Twisted==21.7.0
txaio==21.2.1
typed-ast==1.5.1
typing_extensions==4.0.1
zipp==3.6.0
zope.interface==5.4.0
